Modified UI (Clean)
I added a Taskaler class which will act as our main application which
will initialize and run everything. This is also the class that will
keep track of the global array list of tasks. I also did a simple
wireframe for myself to start wiring up the front end. Take a look and
tell me if there anything wrong with it, besides it not working
completely.

Added:
-> Added A Taskaler.java

Changes:
-> Implemented UI.java to make use of JavaFX
-> Implemented simple UI
-> Added simple event handlers for UI text field