Made changes to UI
I implemented the framework to render a grid calendar. However, the
calendar will always be empty of tasks because I have not implement
logic for that yet. I also wired up the input to execute the command in
controller. Minor cleanup and modifications were made to the README to
better reflect some changes I made on my section.

Changes:
-> Changes to UI.java to dynamically render grid views
-> Cleaned and updated README