Implemented generic list view
I added a new FXML file and an accompanying controller to render a
generic list view. This will be used to render any display that can be
shown as a list. I also cleaned up and added a few method to the
calendar class to allow for monthly, yearly and specified traversal.

Changes:
-> Added listPane FXML and controller class into UI.java
-> Modified calendar class to traverse months, years and to specified
values